added error verifications

+++ client/tecnicofs-client-api.c
@@ -9,7 +9,7 @@
 
 #define MAX_PATH_LEN 30
 
-int sockfd;
+int sockfd, active = 0;
 socklen_t servlen, clilen;
 struct sockaddr_un serv_addr, client_addr;
 
@@ -29,17 +29,22 @@ int setSockAddrUn(char * path, struct sockaddr_un *addr) {
 int tfsCreate(char *filename, char nodeType) {
     char out_buffer[MAX_INPUT_SIZE];
     char in_buffer[MAX_INPUT_SIZE];
+
+    if(!active){
+        printf("tfsCreate: client: doesn't have an active session.\n");
+        return TECNICOFS_ERROR_NO_OPEN_SESSION;
+    }
     
     sprintf(out_buffer,"c %s %c\n",filename,nodeType);
 
     if (sendto(sockfd, out_buffer, strlen(out_buffer)+1, 0, (struct sockaddr *) &serv_addr, servlen) < 0) {
         perror("client: sendto error");
-        exit(EXIT_FAILURE);
+        return TECNICOFS_ERROR_CONNECTION_ERROR;
     }
     
     if(recvfrom(sockfd, in_buffer, sizeof(in_buffer), 0,0,0) < 0){
         perror("client: recvfrom error");
-        exit(EXIT_FAILURE);
+        return TECNICOFS_ERROR_CONNECTION_ERROR;
     }
 
     return atoi(in_buffer);
@@ -48,17 +53,22 @@ int tfsCreate(char *filename, char nodeType) {
 int tfsDelete(char *path) {
     char out_buffer[MAX_INPUT_SIZE];
     char in_buffer[MAX_INPUT_SIZE];
+
+    if(!active){
+        printf("tfsCreate: client: doesn't have an active session.\n");
+        return TECNICOFS_ERROR_NO_OPEN_SESSION;
+    }
     
     sprintf(out_buffer,"d %s\n",path);
 
     if (sendto(sockfd, out_buffer, strlen(out_buffer)+1, 0, (struct sockaddr *) &serv_addr, servlen) < 0) {
         perror("client: sendto error");
-        exit(EXIT_FAILURE);
+        return TECNICOFS_ERROR_CONNECTION_ERROR;
     }
     
     if(recvfrom(sockfd, in_buffer, sizeof(in_buffer), 0,0,0) < 0){
         perror("client: recvfrom error");
-        exit(EXIT_FAILURE);
+        return TECNICOFS_ERROR_CONNECTION_ERROR;
     }
 
     return atoi(in_buffer);
@@ -67,17 +77,22 @@ int tfsDelete(char *path) {
 int tfsMove(char *from, char *to) {
     char out_buffer[MAX_INPUT_SIZE];
     char in_buffer[MAX_INPUT_SIZE];
+
+    if(!active){
+        printf("tfsCreate: client: doesn't have an active session.\n");
+        return TECNICOFS_ERROR_NO_OPEN_SESSION;
+    }
     
     sprintf(out_buffer,"d %s %s\n",from,to);
 
     if (sendto(sockfd, out_buffer, strlen(out_buffer)+1, 0, (struct sockaddr *) &serv_addr, servlen) < 0) {
         perror("client: sendto error");
-        exit(EXIT_FAILURE);
+        return TECNICOFS_ERROR_CONNECTION_ERROR;
     }
     
     if(recvfrom(sockfd, in_buffer, sizeof(in_buffer), 0,0,0) < 0){
         perror("client: recvfrom error");
-        exit(EXIT_FAILURE);
+        return TECNICOFS_ERROR_CONNECTION_ERROR;
     }
 
     return atoi(in_buffer);
@@ -87,35 +102,45 @@ int tfsLookup(char *path) {
     char out_buffer[MAX_INPUT_SIZE];
     char in_buffer[MAX_INPUT_SIZE];
 
+    if(!active){
+        printf("tfsCreate: client: doesn't have an active session.\n");
+        return TECNICOFS_ERROR_NO_OPEN_SESSION;
+    }
+
     sprintf(out_buffer,"l %s\n",path);
 
     if (sendto(sockfd, out_buffer, strlen(out_buffer)+1, 0, (struct sockaddr *) &serv_addr, servlen) < 0) {
         perror("client: sendto error");
-        exit(EXIT_FAILURE);
+        return TECNICOFS_ERROR_CONNECTION_ERROR;
     }
     
     if(recvfrom(sockfd, in_buffer, sizeof(in_buffer), 0,0,0) < 0){
         perror("client: recvfrom error");
-        exit(EXIT_FAILURE);
+        return TECNICOFS_ERROR_CONNECTION_ERROR;
     }
     
     return atoi(in_buffer);
 }
 
 int tfsPrint(char *path) {
-    char out_buffer[MAX_INPUT_SIZE+4];
+    char out_buffer[MAX_INPUT_SIZE];
     char in_buffer[MAX_INPUT_SIZE];
+
+    if(!active){
+        printf("tfsCreate: client: doesn't have an active session.\n");
+        return TECNICOFS_ERROR_NO_OPEN_SESSION;
+    }
 
     sprintf(out_buffer,"p %s\n",path);
 
     if (sendto(sockfd, out_buffer, strlen(out_buffer)+1, 0, (struct sockaddr *) &serv_addr, servlen) < 0) {
         perror("client: sendto error");
-        exit(EXIT_FAILURE);
+        return TECNICOFS_ERROR_CONNECTION_ERROR;
     }
     
     if(recvfrom(sockfd, in_buffer, sizeof(in_buffer), 0,0,0) < 0){
         perror("client: recvfrom error");
-        exit(EXIT_FAILURE);
+        return TECNICOFS_ERROR_CONNECTION_ERROR;
     }
     
     return atoi(in_buffer);
@@ -123,28 +148,41 @@ int tfsPrint(char *path) {
 
 //TODO check errors
 int tfsMount(char * sockPath) {
-    if ((sockfd = socket(AF_UNIX, SOCK_DGRAM, 0)) < 0) {
-        perror("server: can't open socket");
+
+    if(active){
         return TECNICOFS_ERROR_OPEN_SESSION;
+    }
+
+    if ((sockfd = socket(AF_UNIX, SOCK_DGRAM, 0)) < 0) {
+        perror("client: can't open socket");
+        return TECNICOFS_ERROR_OTHER;
     }
 
     clilen = setSockAddrUn(NULL, &client_addr);
 
     if (bind(sockfd, (struct sockaddr *) &client_addr, clilen) < 0) {
-        perror("server: bind error");
+        perror("client: bind error");
         return TECNICOFS_ERROR_OTHER;
     }
     
     servlen = setSockAddrUn(sockPath, &serv_addr);
 
+    active = 1;
+
     return 0;
 }
 
 int tfsUnmount() {
+
+    if(!active){
+        perror("client: doesn't have an active session\n");
+        return TECNICOFS_ERROR_NO_OPEN_SESSION;
+    }
+
     if(close(sockfd) < 0) return FAIL;
 
     if(unlink(client_addr.sun_path) != 0){
-        perror("server: unlink error");
+        perror("client: unlink error");
     }
     
     return SUCCESS;
